Make RDS Gateway RADIUS Client delete graceful when not present

+++ new_dsg_environment/dsg_deploy_scripts/00_clear_dsg_data_from_shm/helper_scripts/Remove_DSG_Data_From_SHM/remote_scripts/Remove_RDS_Gateway_RADIUS_Client_Remote.ps1
@@ -7,5 +7,9 @@
 param(
   $rdsGatewayFqdn
 )
-
-Remove-NpsRadiusClient -Name "$rdsGatewayFqdn"
+if(Get-NpsRadiusClient | Where-Object {$_.Name -eq "$rdsGatewayFqdn"}){
+  Write-Output "Removing RADIUS Client '$rdsGatewayFqdn'"
+  Remove-NpsRadiusClient -Name "$rdsGatewayFqdn"
+} else {
+  Write-Output "No RADIUS Client '$rdsGatewayFqdn' exists"
+}
